working in one page
This prints a grid up to 5k5.  But making it much larger would be
a bit illegible.  The output on the resulting PDF is only on
page 2 for some reason, and the placement is hacky.

Probably what I want to do is one output figure for each *kN,
with different colours for each R.  Then the score (X) axis
can have a reasonable cut value for all inputs.

I think I should write it all in jython/jystilts.